More optimised Editor.js
