+++ src/Components/Components.js
@@ -1,19 +1,7 @@
 import Editor from "./Editor";
 import Hero from "./Hero";
 import Navigation from "./Navigation";
-import TextField from "./TextField";
-import ThreadBox from "./ThreadBox";
 import Tools from "./Tools";
 import LoginCard from "./LoginCard";
-import SideBar from "./SideBar";
 
-export {
-  Editor,
-  Hero,
-  Navigation,
-  TextField,
-  ThreadBox,
-  Tools,
-  LoginCard,
-  SideBar,
-};
+export { Editor, Hero, Navigation, Tools, LoginCard };
