+++ src/Components/Navigation.js
@@ -1,11 +1,16 @@
 import Link from "next/link";
+import Image from "next/image";
 
 import { Logo } from "./Graphics";
 import { useContext } from "react";
 import { FunctionContext } from "@/context/FunctionContext";
+import { AuthContext } from "@/context/AuthContext";
+
+import ProfileImg from "../assets/profile.jpg";
 
 function Navigation() {
   const { setShow } = useContext(FunctionContext);
+  const { user } = useContext(AuthContext);
   return (
     <div className=" h-[7rem] flex justify-between items-center">
       <div className="flex-1">
@@ -22,12 +27,18 @@ function Navigation() {
       </div>
 
       <div className="flex-1 flex justify-end">
-        <button
-          className="bg-[#6F6F6F] px-4 py-1 rounded-xl text-[14px] border-2 border-[#CECECE]"
-          onClick={() => setShow(true)}
-        >
-          Login
-        </button>
+        {user ? (
+          <div className="w-[40px] h-[40px] rounded-full overflow-hidden border-2 border-[#CECECE] cursor-pointer">
+            <Image src={ProfileImg} className="object-cover" />
+          </div>
+        ) : (
+          <button
+            className="bg-[#6F6F6F] px-4 py-1 rounded-xl text-[14px] border-2 border-[#CECECE] cursor-pointer"
+            onClick={() => setShow(true)}
+          >
+            Login
+          </button>
+        )}
       </div>
     </div>
   );
